+++ lib/codec.js
@@ -6,6 +6,9 @@
       , INT = Math.pow(2, 31);
 
   var _ = require('underscore');
+  var protobuf = require('protobufjs');
+  var path = require('path');
+
   var f = require('./functional');
   var utils = require('./utils');
 
@@ -88,6 +91,25 @@
     return _.compose(updateEncOffset(bytebuf), checkedWriteSignedInt(bytebuf), zigZag)(num);
   }
 
+  exports.encodeProtobuf= function(obj,root){
+    return root.encode(obj).finish();
+  }
+
+  exports.encodeWrappedMessage = function(message,typeId){
+    var wrappedMessage={};
+    if(bufferCheck(message)){
+      wrappedMessage.wrappedMessage=message;
+      wrappedMessage.wrappedTypeId=typeId;
+    }else{
+      //it can be extended for the basic data types;
+    }
+    return WrappedMessage.encode(wrappedMessage).finish();
+  }
+
+  // function doEncodeQueryObj(queryObj){
+  //   
+  // }
+
   function zigZag(num) {
     return (num << 1) ^ (num >> 31);
   }
@@ -100,6 +122,11 @@
   var longTooBig = f.validator('must be less than 2^53 (javascript safe integer limitation)',
                                f.lessThan(Math.pow(2, 53)));
   var stringOrNullCheck = f.validator('must be a String or null', stringOrNull);
+  var bufferCheck = f.validator('must be a buffer',isBuffer);
+
+  function isBuffer(x){
+    return Buffer.isBuffer(x);
+  }
 
   function stringOrNull(x) {
     return !f.existy(x) || _.isString(x);
@@ -367,5 +394,20 @@
       return numBytes;
     }
   }
+
+  var WrappedMessage = (function() {
+    var root=protobuf.loadSync(path.join(__dirname+'/protostream/message-wrapping.proto'));
+
+    var wrappedMessage = root.lookupType('org.infinispan.protostream.WrappedMessage');
+
+    return {
+      encode:function(message){
+        return wrappedMessage.encode(message);
+      },
+      decode:function(buf){
+        return wrappedMessage.decode(buf);
+      }
+    }
+  }());
 
 }.call(this));
