+++ lib/protocols.js
@@ -3,8 +3,8 @@
 (function() {
 
   var _ = require('underscore');
-  var protobuf = require('protobufjs');
-  var path = require('path');
+  // var protobuf = require('protobufjs');
+  // var path = require('path');
   
   var f = require('./functional');
   var u = require('./utils');
@@ -772,33 +772,13 @@
       return codec.decodeString();
     }
 
-    function encoderProtobuf(messageInstance){
-      //TODO: validate that arg should be a protobuf messageInstance only 
-      var protobuf = _.compose(encodeWrappedMessage(messageInstance, WrappedMessage), encodeMessageInstance)(messageInstance);
-      return codec.encodeBytesWithLength(protobuf);
+    function encoderProtobuf(message){
+      var typeId = ProtostreamType.lookupProtostreamTypeByName(message.$type.fullName);
+      return f.pipeline(message,f.curry2(codec.encodeProtobuf)(message.$type),f.curry2(codec.encodeWrappedMessage)(typeId),codec.encodeBytesWithLength);
     }
 
     function decoderProtobuf(){
       return f.lift(_.compose(decodeProtobufMessage,decodeWrappedMessage),_.identity)();
-    }
-
-    function encodeMessageInstance(messageInstance){
-      return messageInstance.$type.encode(messageInstance).finish();
-    }
-
-    function encodeWrappedMessage(messageInstance,WrappedMessage){
-      return function (encodedMedia) {
-        var containerTypeName= messageInstance.$type.fullName;
-
-        var containerTypeId = ProtostreamType.lookupProtostreamTypeByName(containerTypeName);
-
-        var wrappedMessage={
-          wrappedTypeId:containerTypeId.protostreamDescriptorId,
-          wrappedMessage:encodedMedia
-        };
-
-        return WrappedMessage.encode(wrappedMessage).finish();
-      }
     }
 
     function decodeProtobufMessage(wrappedMessage){
@@ -1028,27 +1008,12 @@
       lookupProtostreamTypeByName: function (protostreamTypeName) {
         return _.find(protostreamTypes,function(protostreamType){
           return _.isEqual(protostreamType.protostreamTypeName,protostreamTypeName)
-        })
+        }).protostreamDescriptorId;
       },
       lookupProtostreamTypeById: function(protostreamDescriptorId){
         return _.find(protostreamTypes,function(protostreamType){
           return _.isEqual(protostreamType.protostreamDescriptorId,protostreamDescriptorId)
-        })
-      }
-    }
-  }());
-
-  var WrappedMessage = (function() {
-    var root=protobuf.loadSync(path.join(__dirname+'/protostream/message-wrapping.proto'));
-
-    var wrappedMessage = root.lookupType('org.infinispan.protostream.WrappedMessage');
-
-    return {
-      encode:function(message){
-        return wrappedMessage.encode(message);
-      },
-      decode:function(buf){
-        return wrappedMessage.decode(buf);
+        }).protostreamTypeName;
       }
     }
   }());
@@ -1070,7 +1035,7 @@
       }
     }
   }());
-
+  
 
   function Protocol(v, clientOpts) {
     this.version = v;
